\name{acq}
\docType{data}
\alias{acq}
\title{50 exemplary news articles from the Reuters-21578 XML data set with topic acq}
\description{
  This data set holds 50 news articles with additional meta
  information from the Reuters-21578 XML data set.
}
\usage{data(acq)}
\format{A list with 50 text documents.}
\source{Reuters-21578 Text Categorization Collection Distribution 1.0
  (XML format)}
\references{
  Lewis, David (1997) \emph{Reuters-21578 Text Categorization
    Collection Distribution 1.0}.
  \url{http://kdd.ics.uci.edu/databases/reuters21578/reuters21578.html}
}
\keyword{datasets}
