\name{activateCluster}
\alias{activateCluster}
\title{Allow `tm' to Use a Cluster If Available}
\description{
  \pkg{tm} tries to automatically use a (MPI) cluster if
  available on your system. Use this function to manually (re-)activate
  the cluster usage if it has been disabled for some reason.
}
\usage{
activateCluster()
}
\seealso{\code{\link{deactivateCluster}}
}
\author{Ingo Feinerer}
\keyword{file}
