\name{crude}
\docType{data}
\alias{crude}
\title{20 exemplary news articles from the Reuters-21578 XML data set of topic crude}
\description{
  This data set holds 20 news articles with additional meta
  information from the Reuters-21578 XML data set. All documents belong
  to the topic \code{crude} dealing with crude oil.
}
\usage{data("crude")}
\format{A text document collection of 20 text documents.}
\source{Reuters-21578 Text Categorization Collection Distribution 1.0
  (XML format)}
\references{
  Lewis, David (1997) \emph{Reuters-21578 Text Categorization
    Collection Distribution 1.0}.
  \url{http://kdd.ics.uci.edu/databases/reuters21578/reuters21578.html}
}
\examples{
data("crude")
summary(crude)
}
\keyword{datasets}
