\name{deactivateCluster}
\alias{deactivateCluster}
\title{Disallow `tm' to Use a Cluster}
\description{
  \pkg{tm} tries to automatically use a (MPI) cluster if
  available on your system. Use this function to manually forbid
  the cluster usage.
}
\usage{
deactivateCluster()
}
\seealso{\code{\link{activateCluster}}
}
\author{Ingo Feinerer}
\keyword{file}
