\name{Dictionary-class}
\docType{class}
\alias{Dictionary-class}
\title{Dictionary}
\description{
  A class representing a dictionary, i.e., a character vector of terms.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("Dictionary", ...)}.
}
\author{Ingo Feinerer}
\seealso{\code{\link{Dictionary}}}
\keyword{classes}
