\name{eoi-methods}
\docType{methods}
\alias{eoi}
\alias{eoi-methods}
\alias{eoi,DirSource-method}
\alias{eoi,DataframeSource-method}
\alias{eoi,CSVSource-method}
\alias{eoi,URISource-method}
\alias{eoi,ReutersSource-method}
\alias{eoi,GmaneSource-method}
\alias{eoi,VectorSource-method}
\title{Methods for Function eoi in Package `tm'}
\description{
  Methods for function \code{eoi} in package \pkg{tm}..
}
\section{Methods}{
  \describe{
    \item{object = "DataframeSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "DirSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "URISource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "CSVSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "ReutersSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "GmaneSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "VectorSource"}{returns \code{TRUE} if the last item has
      been reached.}
  }
}
\keyword{methods}
