\name{find_freq_terms-methods}
\docType{methods}
\alias{find_freq_terms}
\alias{find_freq_terms-methods}
\alias{find_freq_terms,TermDocMatrix,numeric,numeric-method}
\title{Methods for Function find\_freq\_terms in Package `textmin'}
\description{
Methods for function \code{find_freq_terms} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TermDocMatrix", lowfreq = "numeric", highfreq =
      "numeric"}{Return those terms from \code{object} which occur more or
      equal often than \code{lowfreq} times and less or equal often than
      \code{highfreq} times.
    }
  }
}
\keyword{methods}
