\name{find_hf_terms-methods}
\docType{methods}
\alias{find_hf_terms}
\alias{find_hf_terms-methods}
\alias{find_hf_terms,TermDocMatrix,numeric-method}
\title{Methods for Function find\_hf\_terms in Package `textmin'}
\description{
Methods for function \code{find_hf_terms} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TermDocMatrix", freq = "numeric"}{Return those terms
  from \code{object} which occur more or equal often than \code{freq} times.
}}}
\keyword{methods}
