\name{FunctionGenerator-class}
\docType{class}
\alias{FunctionGenerator-class}
\title{Function Generator}
\description{
  A class representing a function generator.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("FunctionGenerator", ...)}.
}
\section{Extends}{
  Class \code{function}, directly.
}
\author{Ingo Feinerer}
\keyword{classes}
