\name{generics}
\alias{content}
\alias{content<-}
\alias{meta}
\alias{meta<-}
\docType{import}
\title{Imports from package \pkg{NLP}}
\description{
  The generics to access or modify the content or metadata of \R
  objects provided by package \pkg{NLP} are re-exported by package
  \pkg{tm} for convenience.  See
  \code{\link[NLP]{meta}} for more information on the generics and 
  \code{\link{meta.PlainTextDocument}} for the methods for getting and
  setting metadata provided by package \pkg{tm}.
}

