\name{get_elem-methods}
\docType{methods}
\alias{get_elem}
\alias{get_elem-methods}
\alias{get_elem,DirSource-method}
\alias{get_elem,CSVSource-method}
\alias{get_elem,ReutersSource-method}
\alias{get_elem,GmaneRSource-method}
\title{Methods for Function get\_elem in Package `textmin'}
\description{
  Methods for function \code{get_elem} in Package `textmin'.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{returns the element at current position from the source.}
    \item{object = "CSVSource"}{returns the element at current position from the source.}
    \item{object = "ReutersSource"}{returns the element at current position from the source.}
    \item{object = "GmaneRSource"}{returns the element at current position from the source.}
  }
}
\keyword{methods}
