\name{getElem-methods}
\docType{methods}
\alias{getElem}
\alias{getElem-methods}
\alias{getElem,DirSource-method}
\alias{getElem,CSVSource-method}
\alias{getElem,ReutersSource-method}
\alias{getElem,GmaneSource-method}
\title{Methods for Function getElem in Package `textmin'}
\description{
  Methods for function \code{getElem} in Package `textmin'.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{returns the element at current position from the source.}
    \item{object = "CSVSource"}{returns the element at current position from the source.}
    \item{object = "ReutersSource"}{returns the element at current position from the source.}
    \item{object = "GmaneSource"}{returns the element at current position from the source.}
  }
}
\keyword{methods}
