\name{getFilters}
\alias{getFilters}
\title{List Available Filters}
\description{
  List available filters which can be used with \code{\link{tm_filter}}
  and \code{\link{tm_index}}.
}
\usage{
getFilters()
}
\value{
  A character vector with available filters.
}
\author{Ingo Feinerer}
\seealso{
  \code{\link{searchFullText}},
  \code{\link{sFilter}},
  \code{\link{tm_intersect}}.
}
\examples{
getFilters()
}
