\name{getReaders}
\alias{getReaders}
\title{List Available Readers}
\description{
  List available readers.
}
\usage{
getReaders()
}
\value{
  A character vector with available readers.
}
\author{Ingo Feinerer}
\seealso{
  \code{\link{readDOC}},
  \code{\link{readPDF}},
  \code{\link{readPlain}},
  \code{\link{readRCV1}},
  \code{\link{readReut21578XML}},
  \code{\link{readTabular}},
  \code{\link{readXML}}.
}
\examples{
getReaders()
}
