\name{getSources}
\alias{getSources}
\title{List Available Sources}
\description{
  List available sources.
}
\usage{
getSources()
}
\value{
  A character vector with available sources.
}
\author{Ingo Feinerer}
\seealso{
  \code{\link{DataframeSource}},
  \code{\link{DirSource}},
  \code{\link{ReutersSource}},
  \code{\link{URISource}},
  \code{\link{VectorSource}}.
}
\examples{
getSources()
}
