\name{getTransformations}
\alias{getTransformations}
\title{List Available Transformations}
\description{
  List available transformations (mappings) which can be used with
  \code{\link{tm_map}}.
}
\usage{
getTransformations()
}
\value{
  A character vector with available transformations.
}
\seealso{
  \code{\link{removeNumbers}}, \code{\link{removePunctuation}},
  \code{\link{removeWords}}, \code{\link{stemDocument}}, and
  \code{\link{stripWhitespace}}.
}  
\examples{
getTransformations()
}
