\name{\%IN\%-methods}
\alias{\%IN\%}
\docType{methods}
\alias{\%IN\%-methods}
\alias{\%IN\%,TextDocument,TextDocCol-method}
\title{Methods for Function \%IN\% in Package `textmin'}
\description{
  Methods for function \code{\%IN\%} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{x = "TextDocument", y = "TextDocCol"}{Returns \code{TRUE} if
      \code{x} occurs in \code{y}.}
  }
}
\keyword{methods}
