\name{loadDoc-methods}
\docType{methods}
\alias{loadDoc}
\alias{loadDoc-methods}
\alias{loadDoc,PlainTextDocument-method}
\alias{loadDoc,XMLTextDocument-method}
\alias{loadDoc,NewsgroupDocument-method}
\title{Methods for Function loadDoc in Package `textmin'}
\description{
Methods for function \code{loadDoc} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "XMLTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "NewsgroupDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}
}}
\keyword{methods}
