\name{pGetElem-methods}
\docType{methods}
\alias{pGetElem}
\alias{pGetElem-methods}
\alias{pGetElem,DirSource-method}
\alias{pGetElem,DataframeSource-method}
\alias{pGetElem,VectorSource-method}
\title{Methods for Function pGetElem in Package `tm'}
\description{
  Methods for function \code{pGetElem} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{Returns all elements from the source in parallel.}
    \item{object = "DataframeSource"}{Returns all elements from the source in parallel.}
    \item{object = "VectorSource"}{Returns all elements from the source in parallel.}
  }
}
\keyword{methods}
