\name{prescindMeta-methods}
\docType{methods}
\alias{prescindMeta}
\alias{prescindMeta-methods}
\alias{prescindMeta,TextDocCol,character-method}
\title{Methods for Function prescindMetadata in Package `textmin'}
\description{
Methods for function \code{prescindMeta} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", meta = "character"}{Copies the
      \code{meta} data from document slots or (local) document metadata
      of all documents in \code{object} to the (global) \code{DMetaData}
      data frame of \code{object} and returns the whole collection.}
  }
}
\seealso{
  \code{\link{DMetaData}}
}
\keyword{methods}
