\name{RCV1Document-class}
\docType{class}
\alias{RCV1Document-class}
\title{Text document}
\description{
  A class representing a RCV1 XML text document with additional
  information. The XML document itself is represented by inherting from
  \code{XMLTextDocument}.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("RCV1Document", ...)}.
}
\section{Extends}{
Class \code{XMLTextDocument}, directly.
}
\author{Ingo Feinerer}
\keyword{classes}
