\name{removeCitation-methods}
\docType{methods}
\alias{removeCitation}
\alias{removeCitation-methods}
\alias{removeCitation,PlainTextDocument-method}
\title{Methods for Function removeCitation in Package `tm'}
\description{
Methods for function \code{removeCitation} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument"}{Remove citations from e-mail
      messages beginning with \code{>} and return the object.}
  }
}
\keyword{methods}
