\name{removeMultipart-methods}
\docType{methods}
\alias{removeMultipart}
\alias{removeMultipart-methods}
\alias{removeMultipart,PlainTextDocument-method}
\title{Methods for Function removeMultipart in Package `tm'}
\description{
Methods for function \code{removeMultipart} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument"}{Extracts only the text/plain
      parts out of multipart e-mail messages and return the object.}
  }
}
\seealso{
  Use \code{\link{getTransformations}} to list available transformation (mapping) functions.
}
\keyword{methods}
