\name{search_fulltext-methods}
\docType{methods}
\alias{search_fulltext}
\alias{search_fulltext-methods}
\alias{search_fulltext,PlainTextDocument,character-method}
\title{Methods for Function fulltext\_search\_filter in Package `textmin'}
\description{
  Methods for function \code{search_fulltext} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument", pattern = "character"}{return
      \code{TRUE} if the regular expression \code{pattern} matches
      \code{object}'s text corpus.}
  }
}
\keyword{methods}
