\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,TextDocCol-method}
\alias{show,TextRepository-method}
\title{Methods for Function show in Package `tm'}
\description{
Methods for function \code{show} in package \pkg{tm}.
}
\section{Methods}{
\describe{

\item{object = "TextDocCol"}{Prints a short description of \code{object}.}

\item{object = "TextRepository"}{Prints a short description of \code{object}.}
}}
\keyword{methods}
