\name{showMeta-methods}
\docType{methods}
\alias{showMeta}
\alias{showMeta-methods}
\alias{showMeta,TextDocument-method}
\title{Methods for Function showMeta in Package `tm'}
\description{
Methods for function \code{showMeta} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "TextDocument"}{Pretty prints \code{object}'s meta data.}
  }
}
\examples{
data("crude")
showMeta(crude[[1]])
}
\keyword{methods}
