\name{stemDoc-methods}
\docType{methods}
\alias{stemDoc}
\alias{stemDoc-methods}
\alias{stemDoc,PlainTextDocument-method}
\title{Methods for Function stemDoc in Package `textmin'}
\description{
Methods for function \code{stemDoc} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Stems the corpus from \code{object}
  with Porter's stemming algorithm.}
}}
\keyword{methods}
