\name{stepNext-methods}
\docType{methods}
\alias{stepNext}
\alias{stepNext-methods}
\alias{stepNext,Source-method}
\title{Methods for Function stepNext in Package `tm'}
\description{
  Methods for function \code{stepNext} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "Source"}{Jumps to the next element in the source.}
  }
}
\keyword{methods}
