\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[<--methods}
\alias{[[-methods}
\alias{[[<--methods}
\alias{[,TextDocCol-method}
\alias{[<-,TextDocCol-method}
\alias{[[,TextDocCol-method}
\alias{[[<-,TextDocCol-method}
\title{Methods for Subset Functions in Package `tm'}
\description{
  Methods for subset functions \code{[}, \code{[<-}, \code{[[} and
  \code{[[<-} in package \pkg{tm}..
}
\arguments{
  \item{x = "TextDocCol"}{the text document collection to be subsetted
    or modified.}
}
\keyword{methods}
