\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,TextDocCol-method}
\alias{summary,TextRepository-method}
\title{Methods for Function summary in Package `textmin'}
\description{
Methods for function \code{summary} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol"}{Prints a description including information
  on available metadata tags.}

\item{object = "TextRepository"}{Prints a description including information
  on available metadata tags.}
}}
\keyword{methods}
