\name{tm_tolower-methods}
\docType{methods}
\alias{tm_tolower}
\alias{tm_tolower-methods}
\alias{tm_tolower,PlainTextDocument-method}
\title{Methods for Function tm\_tolower in Package `textmin'}
\description{
Methods for function \code{tm_tolower} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", ...}{Converts all
words from \code{object} to lower case and return the object.}
}}
\keyword{methods}
