\name{tm_update-methods}
\docType{methods}
\alias{tm_update}
\alias{tm_update-methods}
\alias{tm_update,TextDocCol,DirSource-method}
\title{Methods for Function tm\_update in Package `textmin'}
\description{
Methods for function \code{tm_update} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", origin = "DirSource"}{Checks whether
      new documents are available in \code{origin}. If, \code{object} is
      updated, i.e., it is augmented with the newly available items.}
  }
}
\keyword{methods}
