\name{tmMap-methods}
\docType{methods}
\alias{tmMap}
\alias{tmMap-methods}
\alias{tmMap,TextDocCol,function-method}
\title{Methods for Function tmMap in Package `textmin'}
\description{
Methods for function \code{tmMap} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", FUN}{Applies \code{FUN} onto each element
  of \code{object}. Note that metadata is automatically passed to
  \code{FUN}. \code{FUN} must be a function which returns a
  \code{TextDocument}.}
}}
\keyword{methods}
