\name{writeCorpus-methods}
\docType{methods}
\alias{writeCorpus}
\alias{writeCorpus-methods}
\alias{writeCorpus,Corpus-method}
\title{Methods for Function writeCorpus in Package `tm'}
\description{
Methods for function \code{writeCorpus} in Package `tm'
}
\section{Methods}{
  \describe{
    \item{object = "Corpus"}{Writes the document collection
      \code{object} containing plain text documents to disk into
      \code{path} using \code{filenames}. In case no manual filenames
      are provided, filenames are automatically generated by using
      \code{object}'s documents' ID strings.}
  }
}
\examples{
data("crude")
\dontrun{writeCorpus(crude, path = ".", filenames = paste(seq_along(crude), ".txt", sep = ""))}
}
\keyword{methods}
