\name{c-methods}
\docType{methods}
\alias{c-methods}
\alias{c,TextDocCol-method}
\alias{c,TextDocument-method}
\title{Methods for Function c in Package `textmin'}
\description{
Methods for function \code{c} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{x = "TextDocCol"}{concatenates several text document collections
      to a single one.}
    \item{x = "TextDocument"}{concatenates several text documents
      to a single text document collection.}
  }
}
\keyword{methods}
