\name{Dictionary}
\docType{methods}
\alias{Dictionary}
\alias{Dictionary,character-method}
\title{Dictionary}
\description{
  Constructs a directory from a character vector.
}
\usage{
\S4method{Dictionary}{character}(object)
}
\arguments{
  \item{object}{a character vector giving the terms for the dictionary.}
}
\value{
  An S4 object of class \code{Dictionary} which extends the class
  \code{character} representing a dictionary.
}
\author{Ingo Feinerer}
\keyword{methods}
