\name{eoi-methods}
\docType{methods}
\alias{eoi}
\alias{eoi-methods}
\alias{eoi,DirSource-method}
\alias{eoi,CSVSource-method}
\alias{eoi,ReutersSource-method}
\alias{eoi,GmaneSource-method}
\title{Methods for Function eoi in Package `textmin'}
\description{
  Methods for function \code{eoi} in Package `textmin'.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "CSVSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "ReutersSource"}{returns \code{TRUE} if the last item has
      been reached.}
    \item{object = "GmaneSource"}{returns \code{TRUE} if the last item has
      been reached.}
  }
}
\keyword{methods}
