\name{findFreqTerms-methods}
\docType{methods}
\alias{findFreqTerms}
\alias{findFreqTerms-methods}
\alias{findFreqTerms,TermDocMatrix,numeric,numeric-method}
\title{Methods for Function findFreqTerms in Package `textmin'}
\description{
Methods for function \code{findFreqTerms} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TermDocMatrix", lowfreq = "numeric", highfreq =
      "numeric"}{Return those terms from \code{object} which occur more or
      equal often than \code{lowfreq} times and less or equal often than
      \code{highfreq} times.
    }
  }
}
\keyword{methods}
