\name{getElem-methods}
\docType{methods}
\alias{getElem}
\alias{getElem-methods}
\alias{getElem,DirSource-method}
\alias{getElem,CSVSource-method}
\alias{getElem,ReutersSource-method}
\alias{getElem,GmaneSource-method}
\alias{getElem,VectorSource-method}
\title{Methods for Function getElem in Package `tm'}
\description{
  Methods for function \code{getElem} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{Returns the element at current position from the source.}
    \item{object = "CSVSource"}{Returns the element at current position from the source.}
    \item{object = "ReutersSource"}{Returns the element at current position from the source.}
    \item{object = "GmaneSource"}{Returns the element at current position from the source.}
    \item{object = "VectorSource"}{Returns the element at current position from the source.}
  }
}
\keyword{methods}
