\name{getTokenizers}
\alias{getTokenizers}
\title{Tokenizers}
\description{
  Predefined tokenizers.
}
\usage{
getTokenizers()
}
\value{
  A character vector with tokenizers provided by package \pkg{tm}.
}
\seealso{
  \code{\link{Boost_tokenizer}}, \code{\link{MC_tokenizer}} and
  \code{\link{scan_tokenizer}}.
}
\examples{
getTokenizers()
}
