\name{getTokenizers}
\alias{getTokenizers}
\title{List Available Tokenizers}
\description{
  List available tokenizers.
}
\usage{
getTokenizers()
}
\value{
  A character vector with available tokenizers.
}
\author{Ingo Feinerer}
\seealso{
  \code{\link{MC_tokenizer}} and \code{\link{scan_tokenizer}}.
}
\examples{
getTokenizers()
}
