\name{getTransformations}
\alias{getTransformations}
\title{Get Available Transformations}
\description{
  Get available transformations (mappings).
}
\usage{
getTransformations()
}
\value{
  A character vector with available transformations.
}
\author{Ingo Feinerer}
\examples{
getTransformations()
}
\keyword{file}
