\name{inspect-methods}
\docType{methods}
\alias{inspect}
\alias{inspect-methods}
\alias{inspect,TextDocCol-method}
\title{Methods for Function inspect in Package `textmin'}
\description{
Methods for function \code{inspect} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol"}{Displays detailed information for
  \code{object}, i.e., it prints all text documents and metadata.}
}}
\keyword{methods}
