\name{loadDoc-methods}
\docType{methods}
\alias{loadDoc}
\alias{loadDoc-methods}
\alias{loadDoc,PlainTextDocument-method}
\alias{loadDoc,XMLTextDocument-method}
\alias{loadDoc,NewsgroupDocument-method}
\alias{loadDoc,StructuredTextDocument-method}
\title{Methods for Function loadDoc in Package `tm'}
\description{
Methods for function \code{loadDoc} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument"}{Loads the text corpus of
      \code{object} from disk into memory and return the new object.}
    \item{object = "XMLTextDocument"}{Loads the text corpus of
      \code{object} from disk into memory and return the new object.}
    \item{object = "NewsgroupDocument"}{Loads the text corpus of
      \code{object} from disk into memory and return the new object.}
    \item{object = "'StructuredTextDocument"}{Loads the text corpus of
      \code{object} from disk into memory and return the new object.}
  }
}
\keyword{methods}
