\name{removeWords-methods}
\docType{methods}
\alias{removeWords}
\alias{removeWords-methods}
\alias{removeWords,PlainTextDocument,character-method}
\title{Methods for Function removeWords in Package `textmin'}
\description{
Methods for function \code{removeWords} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", stopwords = "character"}{Removes all
words which occur in \code{stopwords} from \code{object} and return the object.}
}}
\keyword{methods}
