\name{stepNext-methods}
\docType{methods}
\alias{stepNext}
\alias{stepNext-methods}
\alias{stepNext,DirSource-method}
\alias{stepNext,CSVSource-method}
\alias{stepNext,ReutersSource-method}
\alias{stepNext,GmaneSource-method}
\alias{stepNext,VectorSource-method}
\title{Methods for Function stepNext in Package `tm'}
\description{
  Methods for function \code{stepNext} in package \pkg{tm}..
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{Jumps to the next element in the source.}
    \item{object = "CSVSource"}{Jumps to the next element in the source.}
    \item{object = "ReutersSource"}{Jumps to the next element in the source.}
    \item{object = "GmaneSource"}{Jumps to the next element in the source.}
    \item{object = "VectorSource"}{Jumps to the next element in the source.}
  }
}
\keyword{methods}
