\name{stepNext-methods}
\docType{methods}
\alias{stepNext}
\alias{stepNext-methods}
\alias{stepNext,DirSource-method}
\alias{stepNext,CSVSource-method}
\alias{stepNext,ReutersSource-method}
\alias{stepNext,GmaneSource-method}
\title{Methods for Function stepNext in Package `textmin'}
\description{
  Methods for function \code{stepNext} in Package `textmin'.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{jumps to the next element in the source.}
    \item{object = "CSVSource"}{jumps to the next element in the source.}
    \item{object = "ReutersSource"}{jumps to the next element in the source.}
    \item{object = "GmaneSource"}{jumps to the next element in the source.}
  }
}
\keyword{methods}
