\name{stopwords}
\docType{data}
\alias{stopwords_en}
\title{English Stopwords}
\description{
  This dataset contains very common words that typically want to be
  ignored when processing text.
}
\usage{
   data(stopwords_en)
}
\format{A character vector containing 424 English stopwords.}
\keyword{datasets}
