\name{stripWhitespace-methods}
\docType{methods}
\alias{stripWhitespace}
\alias{stripWhitespace-methods}
\alias{stripWhitespace,PlainTextDocument-method}
\title{Methods for Function stripWhitespace in Package `textmin'}
\description{
Methods for function \code{stripWhitespace} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", ...}{Strips any unnecessary
whitespace from \code{object} and return the object.}
}}
\keyword{methods}
