\name{tmTolower-methods}
\docType{methods}
\alias{tmTolower}
\alias{tmTolower-methods}
\alias{tmTolower,PlainTextDocument-method}
\title{Methods for Function tmTolower in Package `textmin'}
\description{
Methods for function \code{tmTolower} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", ...}{Converts all
words from \code{object} to lower case and return the object.}
}}
\keyword{methods}
