\name{tmUpdate-methods}
\docType{methods}
\alias{tmUpdate}
\alias{tmUpdate-methods}
\alias{tmUpdate,TextDocCol,DirSource-method}
\title{Methods for Function tmUpdate in Package `tm'}
\description{
Methods for function \code{tmUpdate} in package \pkg{tm}.
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", origin = "DirSource", load = FALSE}{Checks whether
      new documents are available in \code{origin}. If, \code{object} is
      updated, i.e., it is augmented with the newly available
      items. \code{load} specifies whether the new items should be
      loaded immediately into memory.}
  }
}
\keyword{methods}
